Improve ccache configuration
For some reason, the ccache on x64 is over 38GB. While I am not sure
why that happened, this aims to fix that by being as explicit as
possible with the configuration.

Also disable ccache compression as this will be later gzip compressed
when uploaded to GCR and we are in no need of compression.

Signed-off-by: rany <rany2@riseup.net>